requirements.txt
aiogram==2.25.1
django==4.2.4
python-dotenv==1.0.0
djangorestframework==3.14.0
requests==2.31.0